move cli tests to cli folder.
