+++ package.json
@@ -73,9 +73,9 @@
     "prettier:check": "prettier --check \"{,**/}*.{js,ts,cjs,mjs,cts,mts,json,md,yml,java,xml}\"",
     "prettier:format": "prettier --write \"{,**/}*.{js,ts,cjs,mjs,cts,mts,json,md,yml,java,xml}\"",
     "pretest": "npm run lint",
-    "test": "npm run mocha-test generators",
+    "test": "npm run mocha-test generators cli",
     "update-snapshot": "mocha --no-insight --no-parallel --updateSnapshot --",
-    "update-snapshots": "mocha test generators --require test/mocha-update-snapshots.config.cjs --no-insight --forbid-only",
+    "update-snapshots": "mocha test generators cli --require test/mocha-update-snapshots.config.cjs --no-insight --forbid-only",
     "watch": "concurrently \"tsc --watch\" \"node ./watch.mjs\""
   },
   "dependencies": {
